Nugit-DS/requirements.txt
# This requirements file lists all dependecies for this project.
# Run 'pip install -r requirements.txt' to install dependencies
#
# E.g:
# numpy==1.7.0
numpy==1.8.1
scipy==0.14.0
simplejson==3.3.0
scikit-learn==0.14.1
pandas==0.14.1
python-nvd3